refactor: Upgrading project to Typescript, refacotring into Routes/Services/Providers/Controller
