+++ package.json
@@ -2,10 +2,9 @@
   "name": "code-challenge",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
   "scripts": {
     "build": "tsc",
-    "start": "npm run build && node dist/index.js",
+    "start": "npm run build && node dist/src/index.js",
     "test": "jest"
   },
   "keywords": [],
