+++ src/providers/database.ts
@@ -1,0 +1,6 @@
+import knex from 'knex';
+import config from '../../knexfile';
+
+const db = knex(config);
+
+export default db;
