+++ src/services/tenantService.ts
@@ -1,0 +1,5 @@
+import db from '../providers/database';
+
+export const getPriceboardsByTenant = async (tenantId: number) => {
+    return db('priceboard').where({ tenant_id: tenantId });
+};
